style: 패키지 이름 오탙자 및 형식 수정
