--- src/main/java/com/trusta_market/walllet_service/domain/enums/RefType.java
@@ -1,5 +1,0 @@
-package com.trusta_market.walllet_service.domain.enums;
-
-public enum RefType {
-	ORDER, PAYMENT
-}
