--- src/main/java/com/trusta_market/walllet_service/domain/enums/WalletStatus.java
@@ -1,5 +1,0 @@
-package com.trusta_market.walllet_service.domain.enums;
-
-public enum WalletStatus {
-	ACTIVE, FROZEN, CLOSED
-}
