--- src/main/java/com/trusta_market/walllet_service/domain/repository/PointTransactionRepository.java
@@ -1,4 +1,0 @@
-package com.trusta_market.walllet_service.domain.repository;
-
-public interface PointTransactionRepository {
-}
