--- src/main/java/com/trusta_market/walllet_service/domain/repository/WalletRepository.java
@@ -1,4 +1,0 @@
-package com.trusta_market.walllet_service.domain.repository;
-
-public interface WalletRepository {
-}
