+++ src/main/java/com/trustamarket/walletservice/domain/entity/PointTransaction.java
@@ -1,10 +1,10 @@
-package com.trusta_market.walllet_service.domain.entity;
+package com.trustamarket.walletservice.domain.entity;
 
 import java.time.Instant;
 import java.util.UUID;
 
-import com.trusta_market.walllet_service.domain.enums.PointTxType;
-import com.trusta_market.walllet_service.domain.enums.RefType;
+import com.trustamarket.walletservice.domain.enums.PointTxType;
+import com.trustamarket.walletservice.domain.enums.RefType;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
