+++ src/main/java/com/trustamarket/walletservice/domain/enums/PointTxType.java
@@ -1,4 +1,4 @@
-package com.trusta_market.walllet_service.domain.enums;
+package com.trustamarket.walletservice.domain.enums;
 
 public enum PointTxType {
 	CHARGE(Direction.INCREASE),
