+++ src/main/java/com/trustamarket/walletservice/domain/enums/RefType.java
@@ -1,0 +1,5 @@
+package com.trustamarket.walletservice.domain.enums;
+
+public enum RefType {
+	ORDER, PAYMENT
+}
