+++ src/main/java/com/trustamarket/walletservice/domain/enums/WalletStatus.java
@@ -1,0 +1,5 @@
+package com.trustamarket.walletservice.domain.enums;
+
+public enum WalletStatus {
+	ACTIVE, FROZEN, CLOSED
+}
