+++ src/main/java/com/trustamarket/walletservice/domain/enums/WalletType.java
@@ -1,4 +1,4 @@
-package com.trusta_market.walllet_service.domain.enums;
+package com.trustamarket.walletservice.domain.enums;
 
 public enum WalletType {
 	USER, SYSTEM_ESCROW, SYSTEM_FEE
