+++ src/main/java/com/trustamarket/walletservice/domain/repository/PointTransactionRepository.java
@@ -1,0 +1,4 @@
+package com.trustamarket.walletservice.domain.repository;
+
+public interface PointTransactionRepository {
+}
